serve: Specify charset=utf-8 in HTTP response Content-Type for text mime types
Absent a charset in the HTTP response, Web browsers will fall back to
legacy encodings for interpreting HTML files, and possibly even other
file types like JavaScript. A standard way to prevent this is

  (HTML) <meta charset="utf-8">
  (JS) <script charset="utf-8" ...>
  (CSS) @charset "utf-8"

On sites with many templates, this may be cumbersome/error-prone to add
everywhere, while configuring the deployment server to send
charset=utf-8 is easy.

This changes `zola serve` to add charset=utf-8 in responses so that this
method can be used without getting mojibake in the local preview.

Possible objection: This might lead people to forget that they need
<meta charset=utf-8> or some server configuration, if it works locally
for them. Answer: this is already the case when people check the result
with `zola build`, since Web browsers normally read local files as
UTF-8.